define bridge network func for each OS
